backend/requirements.txt
# Core FastAPI dependencies with trio support
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Trio async framework (CRITICAL - replaces asyncio)
trio==0.23.1
asks==3.0.0  # HTTP client for trio
trio-asyncio==0.13.0  # Bridge for asyncio compatibility when needed

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Search engines for hybrid search
elasticsearch==8.11.0

# Vector database
infinity-sdk==0.6.0.dev3

# Caching and queues
redis==5.0.1
aioredis==2.0.1

# Document processing core
unstructured[docx,pdf,xlsx]==0.11.2
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
markdownify==0.11.6
tabulate==0.9.0
beautifulsoup4==4.12.2  # HTML/XML parsing
lxml==4.9.3  # XML processing

# Deep Document Understanding (DeepDoc) - ONNX models
onnxruntime==1.16.3
opencv-python==4.8.1.78
pillow==10.1.0
pdfplumber==0.10.3
numpy==1.24.4

# Advanced chunking and RAPTOR
scikit-learn==1.3.2
umap-learn==0.5.4
networkx==3.2.1
graspologic==3.2.0

# NLP and text processing (RAGFlow-specific)
spacy==3.7.2
nltk==3.8.1
tiktoken==0.5.1
sentence-transformers==2.2.2
python-Levenshtein==0.23.0

# RAGFlow-specific performance components (CRITICAL!)
xxhash==3.4.1  # Fast hashing for caching
chardet==5.2.0  # Character encoding detection
readability-lxml==0.8.1  # HTML content extraction
pyclipper==1.3.0  # Geometric operations for layout
shapely==2.0.2  # Geometric computations
editdistance==0.6.2  # String similarity calculations
json-repair==0.7.1  # JSON repair for malformed data

# Advanced ML and NLP libraries
datrie==0.8.2  # Efficient trie-based token storage
cachetools==5.3.3  # Advanced caching strategies
ormsgpack==1.5.0  # High-performance MessagePack serialization

# LLM integration
langchain==0.0.335
langchain-community==0.0.38
langchain-core==0.1.0
yandex-cloud-ml-sdk==0.1.0

# Knowledge Graph
kuzu==0.0.11

# Multi-language text processing
hanziconv==0.3.2  # Traditional/simplified Chinese conversion
cn2an==0.5.22  # Chinese numbers to Arabic
xpinyin==0.7.6  # Chinese to pinyin
roman-numbers==1.0.9  # Roman numerals conversion
word2number==1.1  # Word to number conversion

# Specialized embedding providers
voyageai==0.2.1  # Voyage AI embeddings
bcembedding==0.1.3  # Chinese embeddings
flagembedding==1.2.10  # BGE embeddings
fastembed==0.1.1  # Fast CPU/GPU embeddings

# Chinese LLM providers (optional)
dashscope==1.14.1  # Qwen models
zhipuai==2.0.1  # ChatGLM models

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
click==8.1.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-trio==0.8.0  # Trio-specific testing
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0